v0.2.51 - update to installation instructions for both alphaMELTS for Python and MAGEMinCalc

+++ src/petthermotools/Installation.py
@@ -5,7 +5,116 @@ import sys
 import site
 import sysconfig
 import importlib
+from pathlib import Path
 
+def install_MAGEMinCalc():
+    '''
+    Establish a new julia environment that will be used for any MAGEMin calculations performed through PetThermoTools.
+    MAGEMinCalc and MAGEMin_C are added to this new environment.
+    '''
+    from juliacall import Main as jl
+    env_dir = Path.home() / ".petthermotools_julia_env"
+    env_dir.mkdir(exist_ok=True)
+    jl_env_path = env_dir.as_posix()
+
+    jl.seval(f"""
+        import Pkg
+        Pkg.activate("{jl_env_path}")
+        Pkg.instantiate()
+        """)
+    jl.seval(f"""
+        for pkg in values(Pkg.dependencies())
+            try
+                if pkg.is_direct_dep
+                    Pkg.rm(pkg.name)
+                end
+            catch e
+                @warn "Failed to remove package" pkg=pkg.name exception=e
+            end
+        end
+        Pkg.gc()  # Clean up unused dependencies
+
+        """)
+    jl.seval(f""" 
+        Pkg.Registry.update()
+        Pkg.add("PythonCall")
+             
+        using .Sys
+        if Sys.iswindows()
+            ENV["JULIA_SSL_CA_ROOTS_PATH"] = ""
+        end
+             
+        # Install MAGEMinCalc and dependencies
+        try
+            Pkg.add(url="https://github.com/gleesonm1/MAGEMinCalc.git", rev="v0.5.0")
+        catch e
+            @warn "Failed to install MAGEMinCalc via HTTPS, retrying..." exception=e
+            ENV["JULIA_SSL_CA_ROOTS_PATH"] = ""
+            Pkg.add(url="https://github.com/gleesonm1/MAGEMinCalc.git", rev="v0.5.0")
+        end
+             
+        Pkg.add(name = "MAGEMin_C", version="2.0.6")
+             
+        Pkg.resolve()   
+        Pkg.precompile()
+             
+        println("MAGEMin environment ready at {jl_env_path}")
+        """)
+
+def update_MAGEMinCalc():
+    '''
+    Update MAGEMinCalc and MAGEMin_C to the latest versions that are compatible with PetThermoTools. It is useful to run this function once following each upgrade of PetThermoTools to ensure compatibility.
+    '''
+    from juliacall import Main as jl
+    env_dir = Path.home() / ".petthermotools_julia_env"
+    jl_env_path = env_dir.as_posix()
+
+    jl.seval(f"""
+        import Pkg
+        Pkg.activate("{jl_env_path}")
+
+        # Install MAGEMinCalc and dependencies
+        try
+            Pkg.add(url="https://github.com/gleesonm1/MAGEMinCalc.git", rev="v0.5.0")
+        catch e
+            @warn "Failed to install MAGEMinCalc via HTTPS, retrying..." exception=e
+            ENV["JULIA_SSL_CA_ROOTS_PATH"] = ""
+            Pkg.add(url="https://github.com/gleesonm1/MAGEMinCalc.git", rev="v0.5.0")
+        end
+             
+        Pkg.add(name = "MAGEMin_C", version="2.0.6")
+             
+        Pkg.resolve()   
+        Pkg.precompile()
+        println("MAGEMin environment updated at {jl_env_path}")
+    """)
+
+def test_MAGEMinCalc():
+    '''
+    Test the MAGEMin installation worked! This function should perform a simple fractional crystallization calculation and print the results.
+    '''
+    from juliacall import Main as jl
+    env_dir = Path.home() / ".petthermotools_julia_env"
+    jl_env_path = env_dir.as_posix()
+
+    jl.seval(f"""
+        import Pkg
+        Pkg.activate("{jl_env_path}")
+             
+        using MAGEMinCalc
+
+        comp = Dict("SiO2_Liq" => 47.5, "Al2O3_Liq" => 16.4, "CaO_Liq" => 11.6, "MgO_Liq" => 9.38,
+                    "FeOt_Liq" => 9.16, "K2O_Liq" => 0.329, "Na2O_Liq" => 2.25, "TiO2_Liq" => 2.29, 
+                    "Fe3Fet_Liq" => 0.15, "Cr2O3_Liq" => 0.0, "H2O_Liq" => 0.68)
+
+        Results = MAGEMinCalc.path(comp = comp, T_end_C = 1100.0, dt_C = 2.0, 
+                    P_bar = 1000.0, frac_xtal = true, 
+                    Model = "ig",
+                    find_liquidus = true)
+
+        println(Results["liq1"])
+
+    """)
 
 def install_alphaMELTS(chip="Linux", file_location = None, admin = False):
     '''
@@ -57,7 +166,7 @@ def install_alphaMELTS(chip="Linux", file_location = None, admin = False):
         try:
             urllib.request.urlretrieve(url, zip_path)
         except Exception as e:
-            print(f"Error downloading file: {e}")
+            print(f"Error downloading file: {e}. Please contact Matt Gleeson (gleesonm@berkeley.edu) if this error persists.")
             sys.exit(1)
             return
 
@@ -121,31 +230,57 @@ def install_alphaMELTS(chip="Linux", file_location = None, admin = False):
                     f.write(custom_path)
             except Exception as e:
                 print(f"Error writing to .pth file: {e}")
+                if file_location is None:
+                    custom_path = os.path.join(os.getcwd(), extract_path, zip_path[:-4])
+                else:
+                    custom_path = os.path.join(zip_path[:-4])
+                print(f"""The download of the alphaMELTS for Python files was successful, but these were not added to the Python path (likely due to \n absence of administrator privaledges.
+                    The alphaMELTS for Python files are located at 
+                    {custom_path}.
+                    At the start of each notebook using PetThermoTools please add the following lines of code to add these files to the Python path:
+                    import sys
+                    sys.path.append(r"{custom_path}") """)
                 sys.exit(1)
+        else:
+            if file_location is None:
+                custom_path = os.path.join(os.getcwd(), extract_path, zip_path[:-4])
+            else:
+                custom_path = os.path.join(zip_path[:-4])
+            print(f"""alphaMELTS for Python files have been downloaded and are located at 
+                  {custom_path}.
+                  At the start of each notebook using PetThermoTools please add the following lines of code to add these files to the Python path:
+                  import sys
+                  sys.path.append(r"{custom_path}") """)
 
         return
     
-def remove_alphaMELTS_path():
-    print('Please note this does not remove the previously downloaded alphaMELTS files. That has to be done normally. \n This function simply removes the Python path to those files so that you can install/update to a new version of alphaMELTS for Python.')
-    site_packages_dirs = site.getsitepackages()
+def update_alphaMELTS_path(chip="Linux", file_location = None, admin = False):
+    '''
+    Update the alphaMELTS for Python files to the latest version. \n Please keep an eye of the alphaMELTS Discord Server for information on when a new version of alphaMELTS for Python is published.
+    '''
+    if admin:
+        print('Please note this does not remove the previously downloaded alphaMELTS files. That has to be done manually. \n This function simply removes the Python path to those files so that you can install/update to a new version of alphaMELTS for Python.')
+        site_packages_dirs = site.getsitepackages()
 
-    for dir in site_packages_dirs:
-        print(f"Checking {dir} for .pth files")
-        for file in os.listdir(dir):
-            if file.endswith(".pth"):
-                print(f"Found .pth file: {file}")
+        for dir in site_packages_dirs:
+            print(f"Checking {dir} for .pth files")
+            for file in os.listdir(dir):
+                if file.endswith(".pth"):
+                    print(f"Found .pth file: {file}")
 
-    # Path to your site-packages directory (adjust as needed)
-    site_packages_path = site.getsitepackages()[0]
+        # Path to your site-packages directory (adjust as needed)
+        site_packages_path = site.getsitepackages()[0]
 
-    # Path to the .pth file
-    pth_file_path = os.path.join(site_packages_path, "my_MELTS_path.pth")
+        # Path to the .pth file
+        pth_file_path = os.path.join(site_packages_path, "my_MELTS_path.pth")
 
-    # Remove the .pth file
-    if os.path.exists(pth_file_path):
-        os.remove(pth_file_path)
-        print(f"Removed {pth_file_path}")
-    else:
-        print(f".pth file not found: {pth_file_path}")
+        # Remove the .pth file
+        if os.path.exists(pth_file_path):
+            os.remove(pth_file_path)
+            print(f"Removed {pth_file_path}")
+        else:
+            print(f".pth file not found: {pth_file_path}")
+
+    install_alphaMELTS(chip=chip, file_location = file_location, admin = admin)
 
     return
